Refactor: JSON Parsing 부분 ContactVC에 의존성 주입
